server/requirements-dev.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements-dev.in
#
attrs==20.3.0
    # via pytest
azure-core==1.17.0
    # via
    #   azure-eventgrid
    #   azure-storage-blob
    #   azure-storage-queue
azure-eventgrid==4.5.0
    # via -r requirements.in
azure-storage-blob==12.8.1
    # via -r requirements.in
azure-storage-queue==12.1.6
    # via -r requirements.in
blinker==1.4
    # via sentry-sdk
boto3==1.16.53
    # via -r requirements.in
botocore==1.19.53
    # via
    #   boto3
    #   s3transfer
cachetools==4.2.2
    # via google-auth
certifi==2020.12.5
    # via
    #   minio
    #   msrest
    #   requests
    #   sentry-sdk
cffi==1.14.6
    # via
    #   cryptography
    #   google-crc32c
chardet==4.0.0
    # via requests
click==7.1.2
    # via flask
cryptography==3.4.7
    # via
    #   azure-storage-blob
    #   azure-storage-queue
    #   msal
    #   pyjwt
flake8==3.8.4
    # via -r requirements-dev.in
flask==1.1.2
    # via
    #   -r requirements.in
    #   flask-cors
    #   sentry-sdk
flask-cors==3.0.10
    # via -r requirements.in
google-api-core==2.0.0
    # via
    #   google-cloud-core
    #   google-cloud-storage
google-auth==2.0.1
    # via
    #   google-api-core
    #   google-cloud-core
    #   google-cloud-storage
google-cloud-core==2.0.0
    # via google-cloud-storage
google-cloud-storage==1.42.0
    # via -r requirements.in
google-crc32c==1.1.2
    # via google-resumable-media
google-resumable-media==2.0.0
    # via google-cloud-storage
googleapis-common-protos==1.53.0
    # via google-api-core
greenlet==1.1.1
    # via sqlalchemy
idna==2.10
    # via requests
iniconfig==1.1.1
    # via pytest
isodate==0.6.0
    # via msrest
itsdangerous==1.1.0
    # via flask
jinja2==2.11.2
    # via flask
jmespath==0.10.0
    # via
    #   boto3
    #   botocore
markupsafe==1.1.1
    # via jinja2
mccabe==0.6.1
    # via flake8
minio==7.1.0
    # via -r requirements.in
msal==1.13.0
    # via -r requirements.in
msrest==0.6.21
    # via
    #   azure-eventgrid
    #   azure-storage-blob
    #   azure-storage-queue
mypy==0.790
    # via -r requirements-dev.in
mypy-extensions==0.4.3
    # via mypy
numpy==1.21.2
    # via pandas
oauthlib==3.1.1
    # via requests-oauthlib
packaging==20.8
    # via pytest
pandas==1.3.2
    # via -r requirements.in
pluggy==0.13.1
    # via pytest
protobuf==3.17.3
    # via
    #   google-api-core
    #   googleapis-common-protos
psycopg2-binary==2.8.6
    # via -r requirements.in
py==1.10.0
    # via pytest
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pycodestyle==2.6.0
    # via flake8
pycparser==2.20
    # via cffi
pydantic==1.7.3
    # via -r requirements.in
pyflakes==2.2.0
    # via flake8
pyjwt[crypto]==2.1.0
    # via msal
pyparsing==2.4.7
    # via packaging
pytest==6.2.1
    # via -r requirements-dev.in
python-dateutil==2.8.1
    # via
    #   botocore
    #   pandas
python-json-logger==2.0.1
    # via -r requirements.in
pytz==2021.1
    # via pandas
requests==2.25.1
    # via
    #   -r requirements-dev.in
    #   -r requirements.in
    #   azure-core
    #   google-api-core
    #   google-cloud-storage
    #   msal
    #   msrest
    #   requests-oauthlib
requests-oauthlib==1.3.0
    # via msrest
rsa==4.7.2
    # via google-auth
s3transfer==0.3.4
    # via boto3
sentry-sdk[flask]==0.19.5
    # via -r requirements.in
six==1.15.0
    # via
    #   azure-core
    #   flask-cors
    #   isodate
    #   protobuf
    #   python-dateutil
sqlalchemy==1.4.23
    # via -r requirements.in
toml==0.10.2
    # via pytest
typed-ast==1.4.2
    # via mypy
typing-extensions==3.7.4.3
    # via mypy
urllib3==1.26.2
    # via
    #   botocore
    #   minio
    #   requests
    #   sentry-sdk
werkzeug==1.0.1
    # via flask

# The following packages are considered to be unsafe in a requirements file:
# setuptools
